+++ src/main/java/ua/com/foxminded/university/service/StudentService.java
@@ -17,9 +17,16 @@ public class StudentService {
 	this.studentDao = studentDao;
     }
 
-    public void create(Student createStudent) {
-	// проверить что это уникальный студент (придумать исходя из каких полей)
-	studentDao.create(createStudent);
+    public void create(Student student) {
+	if (studentIsUnique(student)) {
+	    studentDao.create(student);
+	} else {
+	    System.out.println("Student with this name and birthdate already exists, can't create");
+	}
+    }
+
+    public boolean studentIsUnique(Student student) {
+	return studentDao.findByNameAndBirthDate(student.getFirstName(), student.getLastName(), student.getBirthDate()).isEmpty();
     }
 
     public List<Student> findAll() {
@@ -30,12 +37,20 @@ public class StudentService {
 	return studentDao.findById(choice);
     }
 
-    public void update(Student newStudent) {
-	// проверить что это уникальный студент (придумать исходя из каких полей)
-	studentDao.update(newStudent);
+    public void update(Student student) {
+	studentDao.update(student);
     }
 
     public void delete(int id) {
-	studentDao.delete(id);
+	boolean canDelete = idExists(id);
+	if (canDelete) {
+	    studentDao.delete(id);
+	} else {
+	    System.out.println("Can't delete student");
+	}
+    }
+
+    private boolean idExists(int id) {
+	return studentDao.findById(id).isPresent();
     }
 }
